card-cardos-common.c: move common CardOS function to a separate file
