+++ src/libopensc/card-dtrust.c
@@ -28,6 +28,7 @@
 #include <string.h>
 
 #include "asn1.h"
+#include "card-cardos-common.h"
 #include "internal.h"
 
 static const struct sc_card_operations *iso_ops = NULL;
@@ -452,62 +453,6 @@ err:
 }
 
 static int
-_dtrust_compute_shared_value(struct sc_card *card,
-		const u8 *crgram, size_t crgram_len,
-		u8 *out, size_t outlen)
-{
-	int r;
-	struct sc_apdu apdu;
-	u8 *sbuf = NULL;
-
-	if (card == NULL || crgram == NULL || out == NULL) {
-		return SC_ERROR_INVALID_ARGUMENTS;
-	}
-	LOG_FUNC_CALLED(card->ctx);
-	sc_log(card->ctx, "CardOS compute shared value: in-len %" SC_FORMAT_LEN_SIZE_T "u, out-len %" SC_FORMAT_LEN_SIZE_T "u", crgram_len, outlen);
-
-	/* Ensure public key is provided in uncompressed format (indicator byte
-	 * 0x04 followed by X and Y coordinate. */
-	if (crgram_len % 2 == 0 || crgram[0] != 0x04) {
-		LOG_FUNC_RETURN(card->ctx, SC_ERROR_INVALID_ARGUMENTS);
-	}
-
-	/* strip indicator byte */
-	crgram++;
-	crgram_len--;
-
-	sbuf = malloc(crgram_len + 2);
-	if (sbuf == NULL)
-		return SC_ERROR_OUT_OF_MEMORY;
-
-	/* INS: 0x2A  PERFORM SECURITY OPERATION
-	 * P1:  0x80  Resp: Plain value
-	 * P2:  0xA6  Cmd: Control reference template for key agreement */
-	sc_format_apdu(card, &apdu, SC_APDU_CASE_4, 0x2A, 0x80, 0xA6);
-	apdu.resp = out;
-	apdu.resplen = outlen;
-	apdu.le = outlen;
-
-	sbuf[0] = 0x9c; /* context specific ASN.1 tag */
-	sbuf[1] = crgram_len;
-	memcpy(sbuf + 2, crgram, crgram_len);
-	apdu.data = sbuf;
-	apdu.lc = crgram_len + 2;
-	apdu.datalen = crgram_len + 2;
-
-	iso7816_fixup_transceive_length(card, &apdu);
-	r = sc_transmit_apdu(card, &apdu);
-	sc_mem_clear(sbuf, crgram_len + 2);
-	free(sbuf);
-	LOG_TEST_RET(card->ctx, r, "APDU transmit failed");
-
-	if (apdu.sw1 == 0x90 && apdu.sw2 == 0x00)
-		LOG_FUNC_RETURN(card->ctx, (int)apdu.resplen);
-	else
-		LOG_FUNC_RETURN(card->ctx, sc_check_sw(card, apdu.sw1, apdu.sw2));
-}
-
-static int
 dtrust_decipher(struct sc_card *card, const u8 *data,
 		size_t data_len, u8 *out, size_t outlen)
 {
@@ -524,7 +469,7 @@ dtrust_decipher(struct sc_card *card, const u8 *data,
 	case SC_CARD_TYPE_DTRUST_V4_1_MULTI:
 	case SC_CARD_TYPE_DTRUST_V4_1_M100:
 	case SC_CARD_TYPE_DTRUST_V4_4_MULTI:
-		LOG_FUNC_RETURN(card->ctx, _dtrust_compute_shared_value(card, data, data_len, out, outlen));
+		LOG_FUNC_RETURN(card->ctx, cardos_ec_compute_shared_value(card, data, data_len, out, outlen));
 
 	default:
 		return SC_ERROR_NOT_SUPPORTED;
